Check manifest version -> moved to version pkg
