+++ internal/pkg/manifest/manifest.go
@@ -2,7 +2,6 @@ package manifest
 
 import (
 	"fmt"
-	"strings"
 	"sync"
 
 	"go.uber.org/zap"
@@ -27,10 +26,6 @@ type Manifest struct {
 	changed  bool
 	records  *orderedmap.OrderedMap // common map for all: branches, configs and rows manifests
 	lock     *sync.Mutex
-}
-
-type versionInfo struct {
-	Version int `json:"version"`
 }
 
 type Content struct {
@@ -73,19 +68,6 @@ func Load(fs filesystem.Fs, logger *zap.SugaredLogger) (*Manifest, error) {
 	path := filesystem.Join(filesystem.MetadataDir, FileName)
 	if !fs.IsFile(path) {
 		return nil, fmt.Errorf("manifest \"%s\" not found", path)
-	}
-
-	// Read version first
-	version := &versionInfo{}
-	if err := fs.ReadJsonFileTo(path, "manifest", version); err != nil {
-		return nil, err
-	}
-
-	// Validate version, print instructions about migration
-	if warning, err := validateVersion(version.Version); err != nil {
-		return nil, err
-	} else if warning != "" {
-		logger.Warn(`Warning: `, warning)
 	}
 
 	// Read JSON file
@@ -402,22 +384,4 @@ func (m *Manifest) sortRecords() {
 	m.records.Sort(func(a *orderedmap.Pair, b *orderedmap.Pair) bool {
 		return a.Value.(model.ObjectManifest).SortKey(m.SortBy) < b.Value.(model.ObjectManifest).SortKey(m.SortBy)
 	})
-}
-
-func validateVersion(version int) (warning string, err error) {
-	if version < 1 || version > 2 {
-		return "", fmt.Errorf(`unknown version "%d" found in manifest.json`, version)
-	}
-
-	if version == 1 {
-		warning = `
-Your project needs to be migrated to the new version of the Keboola CLI.
-  1. Make sure you have a backup of the current project directory (eg. git commit, git push).
-  2. Then run "kbc pull --force" to overwrite local state.
-  3. Manually check that there are no unexpected changes in the project directory (git diff).
-		`
-		return strings.TrimLeft(warning, "\n"), nil
-	}
-
-	return "", nil
 }
